requirements.txt
openai==1.30.4
langchain==0.2.5
flask==2.3.3
tiktoken==0.7.0
python-dotenv==1.0.1
pytz==2023.3.post1
langchain_community==0.2.5
langchain_openai==0.1.7
pypdfium2==4.30.0
faiss-cpu==1.8.0
pdf2image==1.17.0
pytesseract==0.3.10
openpyxl==3.1.5
pandas==2.2.2
pandas-profiling==3.2.0
fastapi==0.111.0
uvicorn==0.28.1
load_dotenv==0.1.0
python_multipart==0.0.9
streamlit==1.41.1
plotly==6.0.0
pydantic-settings==2.7.1